widen integers; trim an irrelevancy and a mistaken comment
Let's use 16-bit integers, in case we want to run this with a larger
range than 255.

We weren't using PRIMES. (We're not using `factorize_on_system`, either,
but I sentimentally don't feel like cutting it out yet.) We can't
actually get the studies from the distribution-backing hypotheses in a
general setting—I only thought we could because there happens to be a
bijection when divisibility-by-some-n is the only allowable
property. That will change.