pointerless formulation (only allowing a single conj- or dis-junction)
This is verbose and repetitive, but we prefer to cope with copy-pasting
some boilerplate about new hypothesis types into the definition of and
implementations for BasicHypothesis, than to continue the futile
struggle to do it with boxes and traits. If our means of genericity is
just composition of enums, then everything has a known size and is dealt
with directly—we can slap a derivation of Eq, Copy, and Hash on
everything, shove them in a HashMap, and get on with coding our awesome
game instead of fighting the compiler

This is an enormous limitation in the power our system—now, our
so-called "general" hypotheses are a single conjunction or disjunction
of the basics, instead of allowing arbitrarily nested conjunctions or
disjunctions (smuggled inside of a box).

But code is malleable. If we want to redesign everything later, we can.