prelude to defeat
It would be much more elegant to have generic Hypotheses that only
guarantee that the `description` and `predicts_the_hypothesis` methods
are implemented, which take a generic `Study`, and to have a generic
Distribution that builds on all of this. But—as we already fessed up to
earlier in e177a7da—expressing all of this within the constraints of
Rust's type system is really hard (for me), as contrasted to how it
would be trivial in Python (which doesn't really have types in our
sense) and straightforward in Go (whose interfaces aren't fussy about
everything having a known size at compile time).

I'm actually eager to get a result out of this project (playing
something more Zendo-like than our number-guessing game) and don't feel
motivated by the prospect of hours of pain trying to shoehorn trait
objects into being what I want them to be. This being the case, it might
actually make sense (although it breaks my heart) to rename our current
inference module to indicate its specificity to the number-guessing
game, and, in a future commit, to commission a new, largely copy-pasted
module for the triangle science game.

I'm so sorry.