stubby triangle play submodule
Our number of hypotheses is going to be "big" for any interesting
implementation of the "Mezzanine idea", but in the number-guessing game,
our number of studies was very small (default 30!), whereas for triangle
science, we're also going to have a combinatoric blowup in the studies
available for considering. This is likely not only to present
performance challenges later (for which we might turn to some sort of
Monte Carlo approximation rather than computing exact probabilities
deterministically), but will also presently force us to think a bit
about the algorithm to generate them. Maybe it's time to drag in
itertools?