sample entities, for a variable decorative header ... and more to come
The real motive is that we want to be able to randomly sample studies
until we get one with an expected evidential impact sufficiently close
to 1 bit to use (or run out of time trying). But we might as well throw
some life and spontaneity into the decorative display header first.

(There's still at least one more display bug that can cause some
triangles to display fractured. Information theory is simple; drawing
ASCII triangles is fraught with unanticipated subtleties.)